pyproject.toml
[tool.poetry]
name = "python-web-app"
version = "0.1.0"
description = ""
authors = ["Motoki Nishimura <imperturbablefortitude@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.13"
fastapi = "0.81.0"
uvicorn = "0.18.3"
python-multipart = "0.0.6"
pillow = "9.2.0"
onnxruntime = "1.12.1"
numpy = "^1.26"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
